+++ identity/groups.go
@@ -20,7 +20,7 @@ type GroupsAPI struct {
 }
 
 // Create creates a scim group in the Databricks workspace
-func (a GroupsAPI) Create(groupName string, members []string, roles []string, entitlements []string) (group Group, err error) {
+func (a GroupsAPI) Create(groupName string, members []string, roles []string, entitlements []string) (group ScimGroup, err error) {
 	scimGroupRequest := struct {
 		Schemas      []URN           `json:"schemas,omitempty"`
 		DisplayName  string          `json:"displayName,omitempty"`
@@ -50,13 +50,13 @@ func (a GroupsAPI) Create(groupName string, members []string, roles []string, en
 }
 
 // Read reads and returns a Group object via SCIM api
-func (a GroupsAPI) Read(groupID string) (group Group, err error) {
+func (a GroupsAPI) Read(groupID string) (group ScimGroup, err error) {
 	err = a.client.Scim(http.MethodGet, fmt.Sprintf("/preview/scim/v2/Groups/%v", groupID), nil, &group)
 	if err != nil {
 		return
 	}
 	//get inherited groups
-	var groups []Group
+	var groups []ScimGroup
 	for _, inheritedGroup := range group.Groups {
 		inheritedGroupFull, err := a.Read(inheritedGroup.Value)
 		if err != nil {
@@ -80,8 +80,8 @@ func (a GroupsAPI) Filter(filter string) (GroupList, error) {
 }
 
 // GetAdminGroup returns the admin group in a given workspace by fetching with query "displayName+eq+admins"
-func (a GroupsAPI) GetAdminGroup() (Group, error) {
-	var group Group
+func (a GroupsAPI) GetAdminGroup() (ScimGroup, error) {
+	var group ScimGroup
 	var groups GroupList
 	err := a.client.Scim(http.MethodGet, "/preview/scim/v2/Groups", map[string]string{
 		"filter": "displayName+eq+admins",
@@ -143,7 +143,7 @@ func (a GroupsAPI) Delete(groupID string) error {
 }
 
 func (a GroupsAPI) getInheritedAndNonInheritedRoles(
-	group Group, groups []Group) (inherited []RoleListItem,
+	group ScimGroup, groups []ScimGroup) (inherited []RoleListItem,
 	unInherited []RoleListItem) {
 	allRoles := group.Roles
 	var inheritedRoles []RoleListItem